Move echo endpoint to app.py module
It makes sense for the root route (`GET /`) to be in the `app.py` module
(where the FastAPI app instance is located). This will probably be
replaced by current Dragonfly Reporter API version later on in a
separate PR, as the [echo
endpoint](https://github.com/pypi/pypi-observation-api-private-beta?tab=readme-ov-file#get-danger-apiecho)
will most likely be removed in v1.