Add test case for invalid payload
Add a test case that checks that the "kind" in the payload is set to
"malware," but also that "inspector_url" is not present. Normally, the
Observations API requires that the "inspector_url" field be present if
the kind is "is_malware," so let's see if this endpoint enforces that.